Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -3,12 +3,12 @@
 int main() {
     float a,b;
     char c;
-    scanf("%d %d %c",&a,&b,&c);
+    scanf("%f %f %c",&a,&b,&c);
     if( c == '+') {
         int u = a+b;
         printf("%d",u);
     } else if (c == '-') {
-        int u =(int)a-b;
+        int u =a-b;
         printf("%d",u);
     } else if(c == '*') {
         float e = a*b;
